Updated unit tests on MovieDetail class methods, switched movieRating…

+++ MovieDBPI/app/src/main/java/com/mariana/moviedbpi/data/model/database/MovieListDB.kt
@@ -32,7 +32,7 @@ data class MovieDB(
     val releaseDate: String,
 
     @ColumnInfo(name = "user_rating")
-    val userRating: Float,
+    val userRating: Double,
 
     @ColumnInfo(name = "overview")
     val overview: String
